+++ reactor-core/src/main/java/reactor/core/scheduler/PeriodicWorkerTask.java
@@ -16,7 +16,7 @@
 
 package reactor.core.scheduler;
 
-import reactor.core.ContextAware;
+import reactor.util.context.ContextAware;
 import reactor.core.Disposable;
 import reactor.util.annotation.Nullable;
 import reactor.util.context.Context;
